net: mana: add dynamic HWC queue depth with reinit path
After the initial bootstrap channel establishment (queue depth = 1),
query the device for its maximum supported queue depth.  If larger
than the bootstrap depth, tear down the HWC queues and rebuild them
at the reported maximum, then re-establish the channel.

Add mana_hwc_destroy_queues() helper to cleanly tear down CQ (first,
to deregister EQ IRQ), TXQ, RXQ, caller_ctx, and inflight resources.
This helper is used by both the reinit path and can be reused by
future teardown paths.

Validate device-reported dimensions before using them to size DMA
allocations:
  - Enforce protocol minimums (sizeof(gdma_req_hdr/resp_hdr))
  - Check that q_depth * max_msg_size + alignment fits in u32
  - Cap CQ depth to U16_MAX/2 to prevent u16 truncation

The HWC_INIT_DATA_QUEUE_DEPTH field reported by the device is 24 bits
wide, so carry the depth as u32 from mana_hwc_init_event_handler()
through mana_hwc_establish_channel() instead of truncating it to u16
on receipt.  This ensures the clamp and overflow checks above run on
the value the device actually reported: a value above U16_MAX would
otherwise wrap (e.g. 0x10002 -> 2) and silently pass validation as a
small depth, causing reinit at an unintended depth or skipping reinit
entirely.  The narrowing to u16 happens only at the downstream
create_cq() limit, after the value has been clamped to U16_MAX/2.

If reinit fails and the internal teardown of the new channel
succeeded (setup_active cleared), fall back to the bootstrap-depth
channel.  If teardown also failed (setup_active still set), hardware
mappings may still be active — abort channel creation entirely.

In the reinit and fallback paths, mana_hwc_destroy_queues() destroys
the CQ first, which deregisters the EQ from the interrupt dispatch
list (via mana_gd_deregister_irq + list_del_rcu + synchronize_rcu),
ensuring no interrupt handler can access queue memory after it is
freed.

+++ include/net/mana/hw_channel.h
@@ -197,7 +197,7 @@ struct hw_channel_context {
 	u32 max_req_msg_size;
 	u32 max_resp_msg_size;
 
-	u16 hwc_init_q_depth_max;
+	u32 hwc_init_q_depth_max;
 	u32 hwc_init_max_req_msg_size;
 	u32 hwc_init_max_resp_msg_size;
